% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectHeight.R
\name{expectHeight}
\alias{expectHeight}
\title{Predict Height from given Weight}
\usage{
expectHeight(weight = 50)

expectHeight(weight = c(40, 50, 60))
}
\arguments{
\item{weight}{numeric vector}
}
\value{
numeric vector of expected height
}
\description{
predict value with linear model from NHIS dataset
}
\seealso{
\code{\link[=lm]{lm()}}
}
