% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{model}
\alias{model}
\title{Height / Weight regression model}
\format{
R data file
}
\source{
NHISS Korea, https://nhiss.nhis.or.kr/bd/ab/bdaba001cv.do
}
\usage{
model
}
\description{
Simple linear regression model built from NHIS Dataset
}
\keyword{datasets}
